requirements.txt
opencv-python
Pillow
matplotlib
numpy
pandas
pyzbar
streamlit
groq
